+++ web/src/app/data-templates/transformations.ts
@@ -33,6 +33,7 @@ import * as triggerBatchTemplate from './transformations/triggerBatch.json';
 import * as unionTemplate from './transformations/union.json';
 import * as windowTemplate from './transformations/window.json';
 import * as dropcolumnTemplate from './transformations/dropcolumn.json';
+import * as renameColumnTemplate from './transformations/renamecolumn.json';
 
 export const transformations: any = [
     avroTemplate,
@@ -61,7 +62,8 @@ export const transformations: any = [
     triggerBatchTemplate,
     unionTemplate,
     windowTemplate,
-    dropcolumnTemplate
+    dropcolumnTemplate,
+    renameColumnTemplate
 ];
 
 /*********************** */